Refactored and transferred all modals.
