+++ client/components/core/Footer/FooterForm.vue
@@ -65,7 +65,7 @@
     >
       Order a project now
     </UIButton>
-    <SuccessModal id="footer-modal" :visibled="isSuccess" @onClose="handleModalClose" />
+    <SuccessModal id="footer-modal" :visibled="isSuccess" @close="handleModalClose" />
   </form>
 </template>
 
@@ -74,7 +74,7 @@ import { mapActions } from 'vuex'
 import { required, email, maxLength } from 'vuelidate/lib/validators'
 import UIFormCheckboxes from '@/components/shared/UIFormCheckboxes'
 import UIButton from '@/components/shared/UIButton'
-import SuccessModal from '@/components/Modals/success-modal'
+import SuccessModal from '@/components/core/modals/Success'
 import PlaceholderAsterisk from '@/directives/PlaceholderAsterisk'
 
 export default {
