+++ client/components/core/modals/Individuals.vue
