+++ client/components/core/modals/SuccessMessage.vue
@@ -1,5 +1,5 @@
 <template>
-  <section id="success-message" class="success-message" :class="className">
+  <section id="success-message" class="success-message">
     <div class="success-message__conten-wrapper">
       <img src="@/assets/img/common/success-message-hand.svg" alt="Hand" />
       <div class="success-message__text-content">
@@ -19,23 +19,11 @@
 <script>
 export default {
   name: 'SuccessMessage',
-  props: {
-    className: {
-      type: String,
-      default: '',
-    },
-  },
-
-  methods: {
-    close() {
-      this.$emit('close', true)
-    },
-  },
 }
 </script>
 
 <style lang="scss" scoped>
-@import '../../assets/styles/vars';
+@import '../../../assets/styles/vars';
 
 .success-message {
   background-color: $modal-bg-color;
